# Tibble performance - `vec_ptype2()` and `vec_cast()`