--- a/src/cast-dispatch.c
+++ b/src/cast-dispatch.c
@@ -28,6 +28,7 @@ SEXP vec_cast_dispatch(SEXP x,
     default:
       break;
     }
+    break;
 
   case vctrs_type_dataframe:
     switch(class_type(x)) {
@@ -40,6 +41,7 @@ SEXP vec_cast_dispatch(SEXP x,
     default:
       break;
     }
+    break;
 
   case vctrs_type_s3:
     return vec_cast_dispatch2(x, to, x_type, lossy, x_arg, to_arg);
@@ -75,6 +77,7 @@ static SEXP vec_cast_dispatch2(SEXP x, SEXP to,
     default:
       break;
     }
+    break;
 
   case vctrs_class_bare_ordered:
     switch (x_type) {
@@ -93,6 +96,7 @@ static SEXP vec_cast_dispatch2(SEXP x, SEXP to,
     default:
       break;
     }
+    break;
 
   case vctrs_class_bare_tibble:
     switch (x_type) {
@@ -114,6 +118,7 @@ static SEXP vec_cast_dispatch2(SEXP x, SEXP to,
     default:
       break;
     }
+    break;
 
   case vctrs_class_bare_data_frame:
     Rf_errorcall(R_NilValue, "`to` should have been classified as a `vctrs_type_dataframe`");