--- a/src/cast-dispatch.c
+++ b/src/cast-dispatch.c
@@ -29,6 +29,18 @@ SEXP vec_cast_dispatch(SEXP x,
       break;
     }
 
+  case vctrs_type_dataframe:
+    switch(class_type(x)) {
+    case vctrs_class_bare_data_frame:
+      Rf_errorcall(R_NilValue, "`x` should have been classified as a `vctrs_type_dataframe`");
+
+    case vctrs_class_bare_tibble:
+      return df_as_dataframe(x, to, x_arg, to_arg);
+
+    default:
+      break;
+    }
+
   case vctrs_type_s3:
     return vec_cast_dispatch2(x, to, x_type, lossy, x_arg, to_arg);
 
@@ -81,6 +93,30 @@ static SEXP vec_cast_dispatch2(SEXP x, SEXP to,
     default:
       break;
     }
+
+  case vctrs_class_bare_tibble:
+    switch (x_type) {
+    case vctrs_type_dataframe:
+      return df_as_dataframe(x, to, x_arg, to_arg);
+
+    case vctrs_type_s3:
+      switch (class_type(x)) {
+      case vctrs_class_bare_data_frame:
+        Rf_errorcall(R_NilValue, "`x` should have been classified as a `vctrs_type_dataframe`");
+
+      case vctrs_class_bare_tibble:
+        return df_as_dataframe(x, to, x_arg, to_arg);
+
+      default:
+        break;
+      }
+
+    default:
+      break;
+    }
+
+  case vctrs_class_bare_data_frame:
+    Rf_errorcall(R_NilValue, "`to` should have been classified as a `vctrs_type_dataframe`");
 
   default:
     break;