--- a/src/ptype2-dispatch.c
+++ b/src/ptype2-dispatch.c
@@ -33,6 +33,10 @@ SEXP vec_ptype2_dispatch(SEXP x, SEXP y,
   case vctrs_type2_s3_bare_posixlt_bare_posixlt:
     return datetime_datetime_ptype2(x, y);
 
+  case vctrs_type2_s3_dataframe_bare_tibble:
+  case vctrs_type2_s3_bare_tibble_bare_tibble:
+    return tibble_ptype2(x, y, x_arg, y_arg);
+
   default:
     return vec_ptype2_dispatch_s3(x, y, x_arg, y_arg);
   }