--- a/src/type-tibble.c
+++ b/src/type-tibble.c
@@ -3,7 +3,7 @@
 
 // [[ include("vctrs.h") ]]
 SEXP tibble_ptype2(SEXP x, SEXP y, struct vctrs_arg* x_arg, struct vctrs_arg* y_arg) {
-  SEXP out = PROTECT(df_type2(x, y, x_arg, y_arg));
+  SEXP out = PROTECT(df_ptype2(x, y, x_arg, y_arg));
 
   Rf_setAttrib(out, R_ClassSymbol, classes_tibble);