--- a/src/vctrs.h
+++ b/src/vctrs.h
@@ -400,7 +400,7 @@ SEXP vec_cast_dispatch(SEXP x,
                        struct vctrs_arg* x_arg,
                        struct vctrs_arg* to_arg);
 
-SEXP df_type2(SEXP x, SEXP y, struct vctrs_arg* x_arg, struct vctrs_arg* y_arg);
+SEXP df_ptype2(SEXP x, SEXP y, struct vctrs_arg* x_arg, struct vctrs_arg* y_arg);
 SEXP df_as_dataframe(SEXP x, SEXP to, struct vctrs_arg* x_arg, struct vctrs_arg* to_arg);
 
 bool is_data_frame(SEXP x);