--- a/tests/testthat/helper-types.R
+++ b/tests/testthat/helper-types.R
@@ -16,7 +16,8 @@ base_s3_empty_types <- list(
   bare_ordered = new_ordered(),
   bare_date = new_date(),
   bare_posixct = new_datetime(tzone = "UTC"),
-  bare_posixlt = as.POSIXlt(new_datetime(tzone = "UTC"))
+  bare_posixlt = as.POSIXlt(new_datetime(tzone = "UTC")),
+  bare_tibble = tibble::tibble()
 )
 
 proxied_empty_types <- list(