… src/typeof2-s3.c