Closes #592 
Related to tidyverse/tidyr#751

This PR implements C level support for bare tibbles in `vec_ptype2()` and `vec_cast()`.

I also somehow missed some `break;`s in my `vec_cast_dispatch()` switch statements, so this fixes those too.

Basic improvements:

``` r
library(vctrs)
library(tibble)

x <- tibble(x = 1)
```

ptype2

```r
# before
bench::mark(vec_ptype2(x, x))
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(x, x)   10.8µs   12.8µs    75996.    9.41KB     83.7

# after
bench::mark(vec_ptype2(x, x))
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(x, x)   2.92µs   3.53µs   264898.        0B     53.0
```

casting

```r
# before
bench::mark(vec_cast(x, x))
#> # A tibble: 1 x 6
#>   expression          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(x, x)   8.33µs   11.8µs    78897.    8.41KB     15.8

# after
bench::mark(vec_cast(x, x))
#> # A tibble: 1 x 6
#>   expression          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(x, x)   2.27µs    2.8µs   281384.        0B     28.1
```

Rerunning the #592 benchmarks. Finding the common type of tibbles is now as fast as data frames. So `vec_rbind()` is now close to the same speed, but a little slower most of the time. But way better than before!

``` r
library(vctrs)
library(tibble)

df <- data.frame(x = 1, y = 2)
tbl <- as_tibble(df)

lst_df <- rep(list(df), 100000)
lst_tbl <- rep(list(tbl), 100000)

bench::mark(
  vec_ptype_common(!!! lst_df),
  vec_ptype_common(!!! lst_tbl),
  check = FALSE,
  iterations = 20
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression                        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype_common(!!!lst_df)     128ms    255ms      4.05     6.5KB     9.53
#> 2 vec_ptype_common(!!!lst_tbl)    141ms    256ms      3.97    7.27KB     7.55

bench::mark(
  vec_rbind(!!! lst_df),
  vec_rbind(!!! lst_tbl),
  check = FALSE,
  iterations = 20
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(!!!lst_df)     435ms    560ms      1.78    2.68MB     9.09
#> 2 vec_rbind(!!!lst_tbl)    510ms    611ms      1.64    2.67MB     6.90
```

Running the tidyverse/tidyr#751 benchmark Hadley posted at the end against this PR and current master. Nice improvements, but still slower:

``` r
library(tidyr)
n <- 10000

df <- tibble(
  g = 1:n,
  y = rep(list(tibble(x = 1:5)), n)
)

# before
bench::mark(
  unnest(df, y),
  unnest_legacy(df, y),
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest(df, y)         391.9ms  498.1ms      2.00    3.42MB     20.1
#> 2 unnest_legacy(df, y)   52.1ms   58.6ms     16.4     1.25MB     19.0

# after
bench::mark(
  unnest(df, y),
  unnest_legacy(df, y),
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest(df, y)         235.2ms  325.2ms      3.03    3.41MB     13.2
#> 2 unnest_legacy(df, y)   56.9ms   65.3ms     15.1     1.25MB     14.2
```