@DavisVaughan 2020-02-14
Some light benchmarking on `unnest()` seems to indicate that most of the remaining time is split somewhat evenly between a `tidyr::unchop()` line: `map(x, ~new_data_frame(drop_null(.x)))` and the `vec_rbind()` call.

`vec_rbind()` takes around 110ms of the above time, with ~40ms of that being taken up by `vec_match()`. A faster matching function might really help all throughout vctrs, since we use it in a number of places.